Move CalSource to calendar package
