+++ src/main/java/com/penguineering/calmixer/session/SessionController.java
@@ -1,5 +1,6 @@
 package com.penguineering.calmixer.session;
 
+import com.penguineering.calmixer.calendar.CalSource;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.MediaType;
